+++ packages/kokoas-e2e/cypress/e2e/contractPreview2/inputNew.cy.ts
@@ -3,8 +3,7 @@ import { correctInputData, labelMap } from './testData';
 import format from 'date-fns/format';
 import addMonths from 'date-fns/addMonths';
 import { calculateAmount, roundTo } from 'libs';
-import { produce } from 'immer';
-import { IContracts } from 'types';
+import { checkContractAddStatus } from './helpers/checkContractAddStatus';
 
 describe(
   '入力挙動', 
@@ -280,49 +279,16 @@ describe(
     });
 
 
-
-    const checkContractStatus = (mockHasContract: boolean, shouldBeEnabled: boolean) => {
-      cy.intercept('GET', '**/k/v1/records.json?app=231&query=projId*', (req) => {
-        req.continue((res) => {
-          const modifiedBody = produce(res.body as { records: IContracts[] }, (draft) => {
-            if (mockHasContract) {
-              draft.records[0].contractType.value = '契約';
-            } else {
-              draft.records = [];
-            }
-          });
-    
-          res.send({ 
-            body: modifiedBody, 
-          });
-        });
-      }).as('getContracts');
-            
-      cy.wait('@getContracts');
-    
-      cy.getRadiosByValue('追加')
-        .scrollIntoView({ offset: { top: -300, left: 0 } })
-        .should(shouldBeEnabled ? 'be.enabled' : 'be.disabled');
-
-      cy.screenshot();
-    };
-
-
     it.only('本契約がある場合、「追加」は有効状態にになる', () => {
 
-      checkContractStatus(true, true);
-
-     
+      checkContractAddStatus(true, true);
     });
 
     it.only('本契約がない場合、「追加」は無効状態になる', () => {
 
-      checkContractStatus(false, false);
+      checkContractAddStatus(false, false);
 
     });
-
-
-    
   
   },
 );
